refactor: Split translations into separate language files
