+++ packages/chat/app/i18n/translations/en.ts
@@ -1,0 +1,92 @@
+export const en = {
+  common: {
+    welcome: 'Welcome',
+    about: 'About',
+    settings: 'Settings',
+    help: 'Help',
+    history: 'History',
+    clear: 'Clear',
+    save: 'Save',
+    cancel: 'Cancel',
+    delete: 'Delete',
+    confirm: 'Confirm',
+    language: 'Language',
+    appearance: 'Appearance',
+    colorScheme: 'Color Scheme',
+    chooseColorScheme: 'Choose your preferred color scheme',
+    startingFontSize: 'Starting Font Size',
+    theme: 'Theme',
+    chooseTheme: 'Choose your preferred theme',
+    gestures: 'Gestures',
+    shakeAction: 'Shake Action',
+    shakeDescription: 'Choose what happens when you shake your device',
+    none: 'None',
+    clearText: 'Clear Text',
+    clearTextDescription: 'Shake to clear the current text',
+    flashScreen: 'Flash Screen',
+    flashScreenDescription: 'Shake to flash the screen colors',
+    resetDefaults: 'Reset to Defaults',
+    system: 'System',
+    light: 'Light',
+    dark: 'Dark',
+    // Navigation
+    home: 'Home',
+    homeHint: 'Navigate to main text input screen',
+    historyTab: 'History',
+    historyHint: 'View your message history',
+    helpTab: 'Help',
+    helpHint: 'View app instructions and help',
+    settingsTab: 'Settings',
+    settingsHint: 'Adjust app preferences and settings',
+    aboutTab: 'About',
+    aboutHint: 'View app information and credits',
+    mainNavigation: 'Main navigation',
+    // Actions
+    clearHint: 'Erases all current text from the screen',
+    showHistory: 'Show history',
+    showHistoryHint: 'Opens the message history screen',
+    // History Screen
+    searchMessages: 'Search messages',
+    messagesWillAppearHere: 'Messages you create will appear here',
+    clearAllHistory: 'Clear All History',
+    clearHistoryTitle: 'Clear History',
+    clearHistoryConfirmMessage: 'This will permanently delete all messages. Are you sure?',
+    clearAll: 'Clear All',
+    deleteMessage: 'Delete Message',
+    deleteMessageConfirm: 'Are you sure you want to delete this message?',
+    noMessages: 'No messages',
+    // Text Input
+    typeHere: 'Type here',
+    // Help Screen
+    welcomeToOrbiting: 'Welcome to Orbiting',
+    appDescription: 'A simple messaging app for your eyeballs. You can use it to type and display messages to those around you. Display a message loud and clear.',
+    quickStartGuide: 'Quick Start Guide',
+    tapAndType: '✍️ Tap and start typing',
+    tapAndTypeDesc: 'Your message will display loud and clear!',
+    swipeLeft: '⬅️ Swipe Left',
+    swipeLeftDesc: 'To clear the display',
+    swipeRight: '➡️ Swipe Right',
+    swipeRightDesc: 'To view the menu and settings',
+    swipeUp: '⬆️ Swipe Up',
+    swipeUpDesc: 'To display the history',
+    crossPlatform: '📱 Cross-Platform',
+    crossPlatformDesc: 'Works on both desktop and mobile',
+    features: 'Features',
+    autoScaling: 'Auto-scaling text size',
+    autoScalingDesc: 'Text automatically adjusts to fit the screen',
+    messageHistory: 'Message history',
+    messageHistoryDesc: 'Access your previous messages easily',
+    autoSaving: 'Automatic saving',
+    autoSavingDesc: 'All messages are saved automatically',
+    proTips: 'Pro Tips',
+    addToHomescreen: 'Add to Homescreen',
+    addToHomescreenDesc: 'Install the app for quick access anytime',
+    quickLoad: 'Quick Load',
+    quickLoadDesc: 'Tap any message in History to load it',
+    textWrapping: 'Text Wrapping',
+    textWrappingDesc: 'Messages automatically wrap on whitespace',
+    aboutText: 'Orbiting was created by {{author1}} and {{author2}}. It was inspired by the need to communicate with people in the same space, but with different communications needs.',
+    feedbackText: 'Please send us feedback. We want it!',
+    feedbackEmail: 'feedback@orbiting.com'
+  }
+};
