+++ packages/chat/app/i18n/translations/es.ts
@@ -1,0 +1,92 @@
+export const es = {
+  common: {
+    welcome: 'Bienvenido',
+    about: 'Acerca de',
+    settings: 'Ajustes',
+    help: 'Ayuda',
+    history: 'Historial',
+    clear: 'Limpiar',
+    save: 'Guardar',
+    cancel: 'Cancelar',
+    delete: 'Eliminar',
+    confirm: 'Confirmar',
+    language: 'Idioma',
+    appearance: 'Apariencia',
+    colorScheme: 'Esquema de color',
+    chooseColorScheme: 'Elige tu esquema de color preferido',
+    startingFontSize: 'Tamaño de fuente inicial',
+    theme: 'Tema',
+    chooseTheme: 'Elige tu tema preferido',
+    gestures: 'Gestos',
+    shakeAction: 'Acción al agitar',
+    shakeDescription: 'Elige qué sucede cuando agitas tu dispositivo',
+    none: 'Ninguno',
+    clearText: 'Borrar texto',
+    clearTextDescription: 'Agita para borrar el texto actual',
+    flashScreen: 'Destello de pantalla',
+    flashScreenDescription: 'Agita para hacer destellar los colores de la pantalla',
+    resetDefaults: 'Restablecer valores predeterminados',
+    system: 'Sistema',
+    light: 'Claro',
+    dark: 'Oscuro',
+    // Navigation
+    home: 'Inicio',
+    homeHint: 'Ir a la pantalla principal de entrada de texto',
+    historyTab: 'Historial',
+    historyHint: 'Ver historial de mensajes',
+    helpTab: 'Ayuda',
+    helpHint: 'Ver instrucciones y ayuda',
+    settingsTab: 'Ajustes',
+    settingsHint: 'Ajustar preferencias y configuración',
+    aboutTab: 'Acerca de',
+    aboutHint: 'Ver información y créditos',
+    mainNavigation: 'Navegación principal',
+    // Actions
+    clearHint: 'Borra todo el texto actual de la pantalla',
+    showHistory: 'Mostrar historial',
+    showHistoryHint: 'Abre la pantalla de historial de mensajes',
+    // History Screen
+    searchMessages: 'Buscar mensajes',
+    messagesWillAppearHere: 'Los mensajes que crees aparecerán aquí',
+    clearAllHistory: 'Borrar todo el historial',
+    clearHistoryTitle: 'Borrar historial',
+    clearHistoryConfirmMessage: 'Esto eliminará permanentemente todos los mensajes. ¿Estás seguro?',
+    clearAll: 'Borrar todo',
+    deleteMessage: 'Eliminar mensaje',
+    deleteMessageConfirm: '¿Estás seguro de que quieres eliminar este mensaje?',
+    noMessages: 'No hay mensajes',
+    // Text Input
+    typeHere: 'Escribe aquí',
+    // Help Screen
+    welcomeToOrbiting: 'Bienvenido a Orbiting',
+    appDescription: 'Una aplicación de mensajería simple para tus ojos. Puedes usarla para escribir y mostrar mensajes a quienes te rodean. Muestra un mensaje alto y claro.',
+    quickStartGuide: 'Guía de inicio rápido',
+    tapAndType: '✍️ Toca y empieza a escribir',
+    tapAndTypeDesc: '¡Tu mensaje se mostrará alto y claro!',
+    swipeLeft: '⬅️ Desliza a la izquierda',
+    swipeLeftDesc: 'Para limpiar la pantalla',
+    swipeRight: '➡️ Desliza a la derecha',
+    swipeRightDesc: 'Para ver el menú y la configuración',
+    swipeUp: '⬆️ Desliza hacia arriba',
+    swipeUpDesc: 'Para mostrar el historial',
+    crossPlatform: '📱 Multiplataforma',
+    crossPlatformDesc: 'Funciona tanto en escritorio como en móvil',
+    features: 'Características',
+    autoScaling: 'Tamaño de texto automático',
+    autoScalingDesc: 'El texto se ajusta automáticamente para adaptarse a la pantalla',
+    messageHistory: 'Historial de mensajes',
+    messageHistoryDesc: 'Accede fácilmente a tus mensajes anteriores',
+    autoSaving: 'Guardado automático',
+    autoSavingDesc: 'Todos los mensajes se guardan automáticamente',
+    proTips: 'Consejos Pro',
+    addToHomescreen: 'Añadir a la pantalla de inicio',
+    addToHomescreenDesc: 'Instala la aplicación para acceder rápidamente',
+    quickLoad: 'Carga rápida',
+    quickLoadDesc: 'Toca cualquier mensaje en el Historial para cargarlo',
+    textWrapping: 'Ajuste de texto',
+    textWrappingDesc: 'Los mensajes se ajustan automáticamente en los espacios',
+    aboutText: 'Orbiting fue creado por {{author1}} y {{author2}}. Se inspiró en la necesidad de comunicarse con personas en el mismo espacio, pero con diferentes necesidades de comunicación.',
+    feedbackText: '¡Por favor, envíanos tus comentarios. ¡Los queremos!',
+    feedbackEmail: 'feedback@orbiting.com'
+  }
+};
