+++ packages/chat/app/i18n/translations/index.ts
@@ -1,0 +1,7 @@
+import { en } from './en';
+import { es } from './es';
+
+export const resources = {
+  en,
+  es
+};
